Improving the doc. Still that annoying "THE MANIFEST MUST DEFINE A VERSION" bug
Description of the bug:

I had a non-working extension that I could upload to Chrome's webstore

Now I have a version that works locally (see the screenshots folder)

But I just can't update the extension in Chromestore here
https://chrome.google.com/webstore/developer/update

Each time I get this error

```
Upload an extension or app (.zip file)
search-tab-in-chrome-s-omnibox.zip (Server rejected)
An error occurred: Failed to process your item.

The manifest must define a version.
```

But I DO have a version attribute in manifest.json damnit

  "version": "13",

Chrome loads it like a charm locally

And it's not different from what used to work
https://github.com/internaciulo/search-tab-in-chrome-s-omnibox/blob/684b
6fbe22f7030a5517c6e3697db9f836088b6e/manifest.json
{
  ...
  "version": "3",
  ...
}

Even if it is a mistake of mine

Why are so intelorant to mistakes dear computer ? You should know we
are humans, too humans.